- Renaming and moving files to correct folder
