+++ src/lib/src/core/builders/manifest.builder.ts
