+++ src/lib/src/viewer/viewer.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { IiifService } from '../core/iiif-service/iiif-service';
-import { UrlBuilder } from '../core/builders/url-builder';
+import { UrlBuilder } from '../core/builders/url.builder';
 import { Manifest } from '../core/models/manifest';
 import { Options } from '../core/models/options';
-import { ViewerService } from '../core/viewer-service/viewer.service';
+import { ViewerBuilder } from '../core/builders/viewer.builder';
 
 @Component({
   selector: 'mime-viewer',
@@ -13,10 +13,10 @@ import { ViewerService } from '../core/viewer-service/viewer.service';
 export class ViewerComponent implements OnInit {
   @Input() id: string;
   @Input() options: Options;
-  private viewerService: ViewerService;
 
   constructor(
-    private iiifService: IiifService
+    private iiifService: IiifService,
+    private viewerBuilder: ViewerBuilder
   ) { }
 
   ngOnInit(): void {
@@ -28,10 +28,10 @@ export class ViewerComponent implements OnInit {
     if (this.id) {
       this.iiifService.getManifest(UrlBuilder.getManifestUrl(this.id))
         .subscribe((manifest: Manifest) => {
-          this.viewerService = new ViewerService()
+          this.viewerBuilder
             .withOptions(this.options)
             .withTiles(manifest.tileSource)
-            .createViewer();
+            .create();
         });
     }
   }
